Add writeup on linking a library with Autotools
The auto.conf example is modelled on the one in  "21st Century C" by Ben
Klemens. It's a shell script that creates a small C source file (that
prints out some text and includes gmp.h without doing much with it), a
Makefile.am file, and a configure.ac file, which are needed by Autotools
to generate the rest.

The writeup follows by line of thoughts in debugging why compilation
fails and how to find and link a library in an Autotools workflow.